[FLINK-32731][e2e] Fix NameNode uses port 9870 in hadoop3
This closes #23370
